+++ R/supernova-package.R
@@ -1,0 +1,4 @@
+## usethis namespace: start
+#' @importFrom lifecycle deprecated
+## usethis namespace: end
+NULL