+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fix bug where `pairwise()` would not recognize categorical variables if they were created by using
   `factor()` in the formula, e.g. `pairwise(lm(mpg ~ factor(cyl), data = mtcars))`.
 * Fix printing in RMarkdown documents where `supernova()` output was interpreted as a table.
+* Move `estimate-extraction` functions to [`coursekata`](https://github.com/UCLATALL/coursekata-r).
 
 
 # supernova 2.4.4